refactor: 구글 로그인 관련 인터페이스 및 모델 수정
