--- Android/MyApplication/app/src/main/java/com/example/myapplication/data/model/TokenLoginResponse.kt
@@ -1,8 +1,0 @@
-package com.example.myapplication.data.model
-
-data class TokenLoginResponse(
-    val email: String,
-    val type: String,
-    val token: TokenResponse
-)
-
