+++ Android/MyApplication/app/src/main/java/com/example/myapplication/data/model/auth/TokenResponse.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.data.model
+package com.example.myapplication.data.model.auth
 
 data class TokenResponse(
     val accessToken: String,
