+++ Android/MyApplication/app/src/main/java/com/example/myapplication/data/model/common/ApiResponse.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.data.model
+package com.example.myapplication.data.model.common
 data class ApiResponse<T>(
     val body: T?,
     val message: String?
