+++ Android/MyApplication/app/src/main/java/com/example/myapplication/data/model/googlelogin/GoogleLoginByIdTokenRequest.kt
@@ -1,0 +1,9 @@
+package com.example.myapplication.data.model.googlelogin
+
+import com.google.gson.annotations.SerializedName
+
+data class GoogleLoginByIdTokenRequest(
+    @SerializedName("token")
+    val token: String
+)
+
